slightly optimized buildMenu (stupid name btw :P)

+++ smartOrderClient/src/smart/order/client/order/Drink.java
@@ -2,7 +2,7 @@ package smart.order.client.order;
 
 public class Drink extends Menu
 {
-	public Drink(int id, String name, float price)
+	public Drink(int id, String name, double price)
 	{
 		this.id = id;
 		this.name = name;
