+++ smartOrderClient/src/smart/order/client/order/Food.java
@@ -2,7 +2,7 @@ package smart.order.client.order;
 
 public class Food extends Menu
 {
-	public Food(int id, String name, float price)
+	public Food(int id, String name, double price)
 	{
 		this.id = id;
 		this.name = name;
